chore: split mutation log file
