+++ src/query/storages/fuse/src/operations/common/meta/mod.rs
@@ -12,12 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod abort_operation;
-mod meta;
-mod processors;
-mod snapshot_generator;
+mod commit_meta;
+mod mutation_log;
 
-pub use abort_operation::AbortOperation;
-pub use meta::*;
-pub use processors::*;
-pub use snapshot_generator::*;
+pub use commit_meta::*;
+pub use mutation_log::*;
