+++ src/query/storages/fuse/src/operations/common/processors/mod.rs
@@ -13,11 +13,13 @@
 // limitations under the License.
 
 mod sink_commit;
+mod transform_merge_commit_meta;
 mod transform_mutation_aggregator;
 mod transform_serialize_block;
 mod transform_serialize_segment;
 
 pub use sink_commit::CommitSink;
+pub use transform_merge_commit_meta::TransformMergeCommitMeta;
 pub use transform_mutation_aggregator::TableMutationAggregator;
 pub use transform_serialize_block::TransformSerializeBlock;
 pub use transform_serialize_segment::TransformSerializeSegment;
